Revert back the policy usage in the processor
