+++ config/wanda.exs
@@ -7,8 +7,6 @@ config :trento, :messaging, adapter: Trento.Messaging.Adapters.AMQP
 
 config :trento, Trento.Integration.Checks, adapter: Trento.Integration.Checks.Wanda
 
-config :trento, Trento.Integration.Checks.Wanda, policy: Trento.Integration.Checks.Wanda.Policy
-
 config :trento, Trento.Messaging.Publisher, adapter: Trento.Messaging.Adapters.AMQP
 
 config :trento, Trento.Messaging.Adapters.AMQP,
